Integrate macOS macros for multi-platform support
