--- lottie-swift/src/Public/MacOS/AnimationSubview.swift
@@ -1,20 +1,0 @@
-//
-//  AnimationSubview.swift
-//  lottie-swift-iOS
-//
-//  Created by Brandon Withrow on 2/5/19.
-//
-
-import Foundation
-#if os(OSX)
-import AppKit
-
-/// A view that can be added to a keypath of an AnimationView
-public final class AnimationSubview: NSView {
-  
-  var viewLayer: CALayer? {
-    return layer
-  }
-  
-}
-#endif
